Moved planetList into new file
